# Use toolz.unique for possible speedup in ensure_dict