When cytoolz is present, ensure_dict gets a small performance boost in one case.
Otherwise, there is no change in performance.

- [ ] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`