@groutr 2020-02-27
@jrbourbeau I suspected that was the case.  However, in the utils module, I noted a function that import toolz, `dependency_depth` (though it could probably use lru_cache from functools instead of toolz.memoize).

```python
def ensure_dict(d):
    # original implementation
    if type(d) is dict:
        return d
    elif hasattr(d, "dicts"):
        result = {}
        seen = set()
        for dd in d.dicts.values():
            dd_id = id(dd)
            if dd_id not in seen:
                result.update(dd)
                seen.add(dd_id)
        return result
    return dict(d)

def ensure_dict2(d):
    # using toolz.unique
    if type(d) is dict:
        return d
    elif hasattr(d, "dicts"):
        result = {}
        for dd in unique(d.dicts.values(), key=id):
            result.update(dd)
        return result
    return dict(d)
    
def ensure_dict3(d):
    # using toolz.merge and toolz.unique
    if type(d) is dict:
        return d
    elif hasattr(d, "dicts"):
        return merge(unique(d.dicts.values(), key=id))
    return dict(d)
```

With a graph that has 42 values in `dicts.values()`.
```
ensure_dict
168 µs ± 5.94 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
ensure_dict2 (with toolz.unique)
167 µs ± 3.51 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
ensure_dict3 (with toolz.merge and toolz.unique)
165 µs ± 1.74 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

ensure_dict2 (with cytoolz.unique)
160 µs ± 2.98 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
ensure_dict3 (with cytoolz.merge and cytoolz.unique)
151 µs ± 569 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

Perhaps small performance gain isn't worth the trouble of gaining a dependency on toolz.  In that case, I suggest refactoring the function so that the user at least has the option of substituting cytoolz functions. For example
```python
from dask import utils
from cytoolz import unique, merge
utils.unique = cytoolz.unique
utils.merge = cytoolz.merge
```
If `ensure_dict3` form is used, then user can use accelerated unique and merge functions.