@jrbourbeau 2020-03-03
Thanks for providing benchmarks for different implementations of `ensure_dict`! I suspect most users won't manually set the `utils.unique` or `utils.merge` attributes. To make the speedup more broadly available to users, I'd prefer to do something like:

```python
try:
    from cytoolz import merge, unique
except ImportError:
    def merge():
        ...
    
    def unique():
        ...
```

That said, it's not clear to me that the added code complexity is worth the relatively small performance gain. Out of curiosity, does your use case significantly benefit from using the `cytoolz` implementations? 

> I noted a function that import toolz, dependency_depth (though it could probably use lru_cache from functools instead of toolz.memoize)

Good catch. I agree that could be updated to use `functools.lru_cache` instead