@groutr 2020-03-05
@jrbourbeau submitted #5978 to remove the toolz.memoize usage.

Regarding ensure_dict, I thought we were trying to avoid the usage of toolz/cytoolz. I'm not super attached to these changes and agree that if the payoff doesn't cross some threshold, then probably not worth merging.  In my mind, the payoff is:
1. Easier to understand code.  `merge(unique(d.dicts.values(), key=id))` is pretty clear as to what is happening.
2. By expressing it in this way, faster `merge` and `unique` functions could yield a speedup.  `ensure_dict` is called in a loop a couple of places in dask and is always the first step of any optimization. My intuition (which might be mistaken) is that this function probably gets called a lot. A small performance increase per call could yield a larger cumulative performance benefit.  My only use case only shows a small performance benefit.

I felt that these changes would be an overall improvement, but if benefits aren't clear enough, I will not develop this PR further. Perhaps better set changes would be to optionally import cytoolz where ever toolz is already being imported.  This seems to be done in some places, but not in others.