@eriknw 2020-03-05
Just a reminder that `toolz` installs `tlz` as well.  `tlz` looks just like `toolz`, but imports from `cytoolz` if available.  This negates the need for `try-except` around `cytoolz`.