@groutr 2020-03-05
@eriknw thanks for pointing that out. Does dask depend on a new enough version of toolz to make use of that?  `tlz` seems to be in toolz 0.8, but the min version that dask requires is 0.7.3. Perhaps the min version could be bumped to 0.8 (released almost 4 yrs ago)? That would also make the version checking in dask/bag/core unnecessary.