@TomAugspurger 2020-03-06
@groutr can you provide updated benchmarks with the new implementation.

Bumping to toolz>=0.8 and import tlz sounds fine to me, but that should likely be done separately.