@groutr 2020-03-06
@TomAugspurger I agree that using tlz should be done in another PR.

I will work on some more detailed benchmarks.