@jakirkham 2020-03-10
> Just a reminder that `toolz` installs `tlz` as well. `tlz` looks just like `toolz`, but imports from `cytoolz` if available. This negates the need for `try-except` around `cytoolz`.

I did not even know this! 🤯 Thanks Eric! 😄