@mrocklin 2020-03-22
Was this showing up in performance profiles somewhere?  I hesitate to start rewriting toolz code inside of Dask if it's not improving performance of full workloads.  

It's really easy to start microbenchmarking things and improving individual functions without actually improving user experience.  I think that profiling and tuning are really important, but I think that we need to start with more wholesale workloads.  If folks don't mind, I'd prefer to close this PR.