@jrbourbeau 2020-03-23
Closing this PR for now. @groutr if you have a workload where `ensure_dict` is becoming a performance bottleneck, please feel free to reopen